// packages/identity/src/generated/types/identityUser.ts
/**
 * Generated by orval v8.17.0 🍺
 * Do not edit manually.
 * Lemma Identity API
 * OpenAPI spec version: 0.1.0
 */
import type { IdentityUserStatus } from "./identityUserStatus.js";

export interface IdentityUser {
  createdAt: string;
  /**
   * @minLength 1
   * @maxLength 150
   */
  displayName: string;
  /** @minLength 1 */
  email: string;
  /** @pattern ^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-7[0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}$ */
  id: string;
  /** @minLength 1 */
  identityId: string;
  status: IdentityUserStatus;
  updatedAt: string;
}