// packages/identity/src/generated/types/role.ts
/**
 * Generated by orval v8.17.0 🍺
 * Do not edit manually.
 * Lemma Identity API
 * OpenAPI spec version: 0.1.0
 */

export interface Role {
  createdAt: string;
  /**
   * @minLength 1
   * @maxLength 500
   */
  description: string;
  /** @pattern ^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-7[0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}$ */
  id: string;
  isSystem: boolean;
  /** @minLength 1 */
  key: string;
  /**
   * @minLength 1
   * @maxLength 150
   */
  name: string;
  updatedAt: string;
}