// packages/identity/src/generated/types/userRole.ts
/**
 * Generated by orval v8.16.0 🍺
 * Do not edit manually.
 * Lemma Identity API
 * OpenAPI spec version: 0.1.0
 */
import type { UserRoleRoleKey } from "./userRoleRoleKey.js";

export interface UserRole {
  createdAt: string;
  expiresAt: string;
  /** @pattern ^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-7[0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}$ */
  grantedByUserId: string;
  /** @pattern ^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-7[0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}$ */
  roleId: string;
  /** @minLength 1 */
  roleKey: UserRoleRoleKey;
  /** @pattern ^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-7[0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}$ */
  userId: string;
}